% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidity.R
\name{checkValidity}
\alias{checkValidity}
\title{checkValidity}
\usage{
checkValidity(myconstraint, network, onesided = FALSE)
}
\arguments{
\item{myconstraint}{the formula}

\item{network}{the network object}

\item{onesided}{TRUE for alias, default FALSE}
}
\value{
a message (TRUE if valid otherwise the error message)
}
\description{
checkValidity
}
