% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrEditorServer.R
\name{constrEditorServer}
\alias{constrEditorServer}
\title{constrEditorServer}
\usage{
constrEditorServer(id, network, slot, tab)
}
\arguments{
\item{id}{the id of the ui}

\item{network}{the current network}

\item{slot}{either components or aliases}

\item{tab}{selected tab}
}
\value{
an updated network
}
\description{
constrEditorServer
}
