% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrEditorUI.R
\name{constrEditorUI}
\alias{constrEditorUI}
\title{constrEditorUI
Editor of constrEditorUI}
\usage{
constrEditorUI(id)
}
\arguments{
\item{id}{the id of ui}
}
\description{
constrEditorUI
Editor of constrEditorUI
}
