% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertConstr2idConstr.R
\name{convertConstr2idConstr}
\alias{convertConstr2idConstr}
\title{convertConstr2idConstr}
\usage{
convertConstr2idConstr(constraint, dictionary)
}
\arguments{
\item{constraint}{the constraint}

\item{dictionary}{the dictionary}
}
\value{
a formula in id wording
}
\description{
convertConstr2idConstr
}
