% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertidConstr2Constr.R
\name{convertidConstr2Constr}
\alias{convertidConstr2Constr}
\title{convertidConstr2Constr}
\usage{
convertidConstr2Constr(constraint, dictionary)
}
\arguments{
\item{constraint}{the constraint with id}

\item{dictionary}{the dictionary}
}
\value{
a formula in dictionary
}
\description{
convertidConstr2Constr
}
