% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyAliases.R
\name{createEmptyAliases}
\alias{createEmptyAliases}
\title{createEmptyAliases}
\usage{
createEmptyAliases()
}
\value{
an empty aliases data.frame
}
\description{
createEmptyAliases
}
\examples{
createEmptyAliases()

}
