% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyComponents.R
\name{createEmptyComponents}
\alias{createEmptyComponents}
\title{createEmptyComponents
this function creates an empty components data.reme}
\usage{
createEmptyComponents()
}
\value{
an empty component data frame
}
\description{
createEmptyComponents
this function creates an empty components data.reme
}
