% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyConstraints.R
\name{createEmptyConstraints}
\alias{createEmptyConstraints}
\title{createEmptyConstraints
this function creates an empty constraints data.frame}
\usage{
createEmptyConstraints()
}
\value{
an empty component data frame
}
\description{
createEmptyConstraints
this function creates an empty constraints data.frame
}
