% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyFluxes.R
\name{createEmptyFluxes}
\alias{createEmptyFluxes}
\title{createEmptyFluxes
creates an empty fluxes data frame}
\usage{
createEmptyFluxes()
}
\value{
a data frame
}
\description{
createEmptyFluxes
creates an empty fluxes data frame
}
\examples{
createEmptyFluxes()

}
