% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyMetaObs.R
\name{createEmptyMetaObs}
\alias{createEmptyMetaObs}
\title{createEmptyMetaObs
creates an empty meta obs data frame}
\usage{
createEmptyMetaObs()
}
\value{
a data frame
}
\description{
createEmptyMetaObs
creates an empty meta obs data frame
}
