% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyNetwork.R
\name{createEmptyNetwork}
\alias{createEmptyNetwork}
\title{createEmptyNetwork}
\usage{
createEmptyNetwork()
}
\value{
a reactive network with empty components
}
\description{
createEmptyNetwork
}
\examples{
createEmptyNetwork()
}
