% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEmptyTimeLine.R
\name{createEmptyTimeLine}
\alias{createEmptyTimeLine}
\title{createEmptyTimeLine}
\usage{
createEmptyTimeLine()
}
\value{
an empty timeline datafame
}
\description{
createEmptyTimeLine
}
\examples{
createEmptyTimeLine

}
