% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileInteractionUI.R
\name{fileInteractionUI}
\alias{fileInteractionUI}
\title{fileInteractionUI
ui of the file interaction visual editor}
\usage{
fileInteractionUI(id)
}
\arguments{
\item{id}{the id of the ui}
}
\value{
nothing
}
\description{
fileInteractionUI
ui of the file interaction visual editor
}
