% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_with_rownames.R
\name{filter_with_rownames}
\alias{filter_with_rownames}
\title{filter_with_rownames}
\usage{
filter_with_rownames(data, ...)
}
\arguments{
\item{data}{the data ti filter}

\item{...}{argument sent to filter}
}
\value{
filtered data with kept rownames
}
\description{
filter_with_rownames
}
\examples{
filter_with_rownames(iris, Species == "versicolor")

}
