% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDictionary.R
\name{generateDictionary}
\alias{generateDictionary}
\title{generateDictionary}
\usage{
generateDictionary(components, fluxes, observations, metaobs, aliases = NULL)
}
\arguments{
\item{components}{the components}

\item{fluxes}{the fluxes}

\item{observations}{the observations}

\item{metaobs}{meta information of observation}

\item{aliases}{facultative aliases}
}
\value{
a dictionary
}
\description{
generateDictionary
}
