% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSymbolicEnvir.R
\name{generateSymbolicEnvir}
\alias{generateSymbolicEnvir}
\title{generateSymbolicEnvir}
\usage{
generateSymbolicEnvir(network)
}
\arguments{
\item{network}{the network object}
}
\value{
an environment with symbolic environment
}
\description{
generateSymbolicEnvir
}
