% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstraintWord.R
\name{getConstraintWord}
\alias{getConstraintWord}
\title{getConstraintWord}
\usage{
getConstraintWord(constraint)
}
\arguments{
\item{constraint}{the equation of the constraint}
}
\value{
a vector with words of the formula
}
\description{
getConstraintWord
}
