% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_tol.R
\name{ident_tol}
\alias{ident_tol}
\title{ident_tol
compares x and y and returns TRUE or FALSE if x and y are equal given
tolerance}
\usage{
ident_tol(x, y, tolerance = 0.001)
}
\arguments{
\item{x}{object 1}

\item{y}{object 2}

\item{tolerance}{numerical toleance}
}
\value{
a logical
}
\description{
ident_tol
compares x and y and returns TRUE or FALSE if x and y are equal given
tolerance
}
\examples{
require(dplyr)
ident_tol(tibble(iris),
    tibble(iris |>
      mutate(Petal.Length = Petal.Length + .1)),
         tolerance = 0.2)
         
ident_tol(tibble(iris),
    tibble(iris |>
      mutate(Petal.Length = Petal.Length + .1)),
         tolerance = 0.01)
}
