% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchConstructor.R
\name{launchConstructor}
\alias{launchConstructor}
\title{launch the Shiny App}
\usage{
launchConstructor(...)
}
\arguments{
\item{...}{options to shinyApp}
}
\value{
nothing
}
\description{
launch the Shiny App
}
