% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRCaNfile.R
\name{loadRCaNfile}
\alias{loadRCaNfile}
\title{loadRCaNfile}
\usage{
loadRCaNfile(datapath, modelname)
}
\arguments{
\item{datapath}{the path to the model}

\item{modelname}{the name of the model}
}
\value{
a list with the slots of the network
}
\description{
loadRCaNfile
}
