% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCaNconstructorUI.R
\name{RCaNconstructorUI}
\alias{RCaNconstructorUI}
\title{RCaNconstructorUI
function that returns the ui of the the Constructor}
\usage{
RCaNconstructorUI()
}
\value{
an ui
}
\description{
RCaNconstructorUI
function that returns the ui of the the Constructor
}
