% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabConstrServer.R
\name{tabConstrServer}
\alias{tabConstrServer}
\title{tabConstrServer}
\usage{
tabConstrServer(id, network, tab)
}
\arguments{
\item{id}{the id of the ui}

\item{network}{the current network}

\item{tab}{selected tab}
}
\value{
an updated network
}
\description{
tabConstrServer
}
