% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabConstrUI.R
\name{tabConstrUI}
\alias{tabConstrUI}
\title{tabConstrUI
Visualize tables}
\usage{
tabConstrUI(id)
}
\arguments{
\item{id}{the id of ui}

\item{title}{the title to display}
}
\description{
tabConstrUI
Visualize tables
}
