% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableEditorUI.R
\name{tabEditorUI}
\alias{tabEditorUI}
\title{tabEditorUI
Editor of tables}
\usage{
tabEditorUI(id, title)
}
\arguments{
\item{id}{the id of ui}

\item{title}{the title to display}
}
\description{
tabEditorUI
Editor of tables
}
