% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableObsServer.R
\name{tableObsServer}
\alias{tableObsServer}
\title{tableObsServer}
\usage{
tableObsServer(id, network, tab)
}
\arguments{
\item{id}{the id of the ui}

\item{network}{the current network}

\item{tab}{selected tab}
}
\value{
a data.frame observation
}
\description{
tableObsServer
}
