% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableObsUI.R
\name{tableObsUI}
\alias{tableObsUI}
\title{tableObsUI
Editor of tables}
\usage{
tableObsUI(id, title)
}
\arguments{
\item{id}{the id of ui}

\item{title}{the tital to display}
}
\description{
tableObsUI
Editor of tables
}
