% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabObsMetaServer.R
\name{tabObsMetaServer}
\alias{tabObsMetaServer}
\title{tabObsMetaServer}
\usage{
tabObsMetaServer(id, network, tab)
}
\arguments{
\item{id}{the id of the ui}

\item{network}{the current network}

\item{tab}{selected tab}
}
\value{
an updated network
}
\description{
tabObsMetaServer
}
