% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineServer.R
\name{timeLineServer}
\alias{timeLineServer}
\title{timeLineServer}
\usage{
timeLineServer(id, timeline)
}
\arguments{
\item{id}{the id of the ui}

\item{timeline}{the timeline}
}
\value{
an updated network
}
\description{
timeLineServer
}
