% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeLineUI.R
\name{timeLineUI}
\alias{timeLineUI}
\title{timeLineUI
ui of the file timeline editorœ}
\usage{
timeLineUI(id)
}
\arguments{
\item{id}{the id of the ui}
}
\value{
nothing
}
\description{
timeLineUI
ui of the file timeline editorœ
}
