% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetworkServer.R
\name{visNetworkServer}
\alias{visNetworkServer}
\title{visNetworkServer}
\usage{
visNetworkServer(id, datanet)
}
\arguments{
\item{id}{the id of the visnetworkUI}

\item{datanet}{the data of the graph}
}
\value{
updated newtork
}
\description{
visNetworkServer
}
