% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetworkUI.R
\name{visNetworkUI}
\alias{visNetworkUI}
\title{visNetworkUI
ui of the network visual editor}
\usage{
visNetworkUI(id)
}
\arguments{
\item{id}{the id of the ui}
}
\value{
nothing
}
\description{
visNetworkUI
ui of the network visual editor
}
