% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRCaNfile.R
\name{writeRCaNfile}
\alias{writeRCaNfile}
\title{writeRCaNfile}
\usage{
writeRCaNfile(modelfile, filenewnetwork, param, input, orig)
}
\arguments{
\item{modelfile}{the workbook to write}

\item{filenewnetwork}{the description of the network}

\item{param}{the parameter of the info}

\item{input}{the input for which to retrieve info}

\item{orig}{the original file}
}
\value{
an updated modefile
}
\description{
writeRCaNfile
}
