% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTimeLine.R
\name{writeTimeLine}
\alias{writeTimeLine}
\title{writeTimeLine}
\usage{
writeTimeLine(type, new, old)
}
\arguments{
\item{type}{}

\item{new}{}

\item{old}{}
}
\value{
a string describing the change
}
\description{
writeTimeLine
}
\examples{
writeTimeLine("aliases", data.frame(Alias = "toto", Formula = 33), NULL)

}
