% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileInteractionServer.R
\name{fileInteractionServer}
\alias{fileInteractionServer}
\title{fileInteractionServer}
\usage{
fileInteractionServer(id, network, timeline)
}
\arguments{
\item{id}{the id of the ui}

\item{network}{the network}

\item{timeline}{the timeline}
}
\value{
a updated network
}
\description{
fileInteractionServer
}
