% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableEditorServer.R
\name{tableEditorServer}
\alias{tableEditorServer}
\title{tableEditorServer}
\usage{
tableEditorServer(id, network, slot, tab)
}
\arguments{
\item{id}{the id of the ui}

\item{network}{the current network}

\item{slot}{either fluxes or components}

\item{tab}{selected tab}
}
\value{
an updated network
}
\description{
tableEditorServer
}
