% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNetworkServer.R
\name{visNetworkServer}
\alias{visNetworkServer}
\title{visNetworkServer}
\usage{
visNetworkServer(id, network, tab)
}
\arguments{
\item{id}{the id of the visnetworkUI}

\item{network}{the data of the graph}

\item{tab}{selected tab}
}
\value{
updated newtork
}
\description{
visNetworkServer
}
