% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRCaNfile.R
\name{writeRCaNfile}
\alias{writeRCaNfile}
\title{writeRCaNfile}
\usage{
writeRCaNfile(modelfile, filenewnetwork)
}
\arguments{
\item{modelfile}{the workbook to write}

\item{filenewnetwork}{the description of the network}
}
\value{
an updated modefile
}
\description{
writeRCaNfile
}
